Light driver Ok (high mode without transistor on GND), add light for node 6 and 7 on collegio-node app, remove pressure for node 5 and 3 on collegio-node app, add validation for pressure and temperature on collegio-node app, doubled sending rate for node 5 of collegio-node app.

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/sensors/Light.java
@@ -30,9 +30,8 @@ public class Light
 { 
 	/**
 	 * Reads light value from the BH1620FVC sensor.	 
-	 * @return measured light express in mLx.
+	 * @return measured light express in Lx (Lux).
 	 */
-	//TODO: c'è un motivo per ritornare un int e non uno short ?
-	public static native int getValue();
+	public static native short getValue();
 
 }
